+++ CHANGES.md
@@ -15,6 +15,10 @@
     - @miosakuma
 - [ADD] 転送フィルター機能を追加する
     - @szktty
+- [ADD]
+  - scalability mode に対応する
+  - VP9 / AV1 のサイマルキャストに対応可能になる
+  - @szktty
 
 ## 2023.1.0